backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Machine Learning
scikit-learn==1.3.2
torch==2.1.1
transformers==4.36.0
xgboost==2.0.3
numpy==1.24.4
pandas==2.1.4

# Data Processing
datasets==3.6.0
huggingface-hub==0.20.2
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.0

# API & Utilities
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Production
gunicorn==21.2.0
supervisor==4.2.5
